[cucumber-react] Extract Cucumber-React to separate sub project
