+++ SUMMARY.md
@@ -103,6 +103,7 @@
 
 * [Official Extensions](#)
   * [Gherkin-Lint](gherkin-lint/README.md)
+  * [Cucumber-React](cucumber-react/README.md)
   * [CLI automation - Aruba](#)
   * [Ruby on Rails](#)
   * [Spring Framework](#)
