+++ cucumber-react/.travis.yml
@@ -1,0 +1,4 @@
+# manyrepo: cucumber/gherkin-lint-javascript
+language: node_js
+node_js:
+  - "6"
