+++ cucumber-react/README.md
@@ -1,0 +1,8 @@
+# Cucumber-React
+
+React components for rendering Gherkin documents and Cucumber results.
+
+This library also contains a [Redux](http://redux.js.org/) reducer that
+builds an [Immutable.js](https://facebook.github.io/immutable-js/) state using
+[Event Protocol](../event-protocol) events as
+[Redux actions](http://redux.js.org/docs/basics/Actions.html)
