+++ cucumber-react/cucumber.js
@@ -1,0 +1,3 @@
+module.exports = {
+  default: '--compiler js:babel-core/register --snippet-interface promise'
+}
