+++ cucumber-react/lib/index.js
@@ -1,0 +1,4 @@
+import * as CucumberReact from './cucumber_react'
+import reducer from './reducer'
+
+export { CucumberReact, reducer }
