+++ cucumber-react/lib/reducer.js
