+++ cucumber-react/package.json
@@ -1,0 +1,49 @@
+{
+  "name": "cucumber-react",
+  "version": "1.0.0",
+  "description": "React component for Cucumber",
+  "main": "lib/index.js",
+  "engines": {
+    "node": ">=6.5.0"
+  },
+  "scripts": {
+    "test": "npm run eslint && npm run mocha",
+    "eslint": "eslint lib test",
+    "mocha": "mocha"
+  },
+  "repository": {
+    "type": "git",
+    "url": "git://github.com/cucumber/cucumber-react.git"
+  },
+  "keywords": [
+    "cucumber",
+    "gherkin",
+    "react"
+  ],
+  "author": "Cucumber Ltd",
+  "license": "MIT",
+  "bugs": {
+    "url": "https://github.com/cucumber/cucumber/issues"
+  },
+  "homepage": "https://github.com/cucumber/cucumber-react#readme",
+  "devDependencies": {
+    "babel-cli": "^6.11.4",
+    "babel-loader": "^6.2.5",
+    "babel-preset-es2015": "^6.13.2",
+    "babel-preset-react": "^6.11.1",
+    "enzyme": "^2.4.1",
+    "eslint": "^3.5.0",
+    "eslint-config-eslint": "^3.0.0",
+    "eslint-plugin-node": "^2.0.0",
+    "eslint-plugin-react": "^6.2.2",
+    "json-loader": "^0.5.4",
+    "mocha": "^3.0.2",
+    "react": "^15.3.1",
+    "react-addons-test-utils": "^15.3.1",
+    "react-dom": "^15.3.1",
+    "webpack": "^1.13.2"
+  },
+  "dependencies": {
+    "gherkin": "^4.0.0"
+  }
+}
