+++ cucumber-react/test/mocha.opts
@@ -1,0 +1,2 @@
+--compilers js:babel-core/register
+--recursive
