--- html-formatter/nodejs/lib/cucumber_react/README.md
@@ -1,2 +1,0 @@
-Cucumber-React uses the [Gherkin AST]() as its data model (in Immutable.js format).
-There is an addition: Attachments!!
