+++ html-formatter/nodejs/package.json
@@ -33,7 +33,6 @@
   },
   "homepage": "https://github.com/cucumber/cucumber-html-formatter-nodejs#readme",
   "devDependencies": {
-    "babel": "^6.5.2",
     "babel-cli": "^6.11.4",
     "babel-loader": "^6.2.5",
     "babel-preset-es2015": "^6.13.2",
@@ -52,6 +51,7 @@
     "webpack": "^1.13.2"
   },
   "dependencies": {
+    "cucumber-react": "^1.0.0",
     "debug": "^2.2.0",
     "event-stream": "^3.3.4",
     "eventsource": "^0.2.1",
