+++ html-formatter/nodejs/webpack.config.js
@@ -8,7 +8,6 @@ module.exports = {
     loaders: [
       {
         test: /\.jsx?$/,
-        exclude: /node_modules/,
         loader: 'babel-loader'
       },
       {
